added filter + pulse generator + UTs; some fixes needed

+++ include/SupportedFormats.h
@@ -4,9 +4,9 @@
 
 #pragma once
 
-constexpr unsigned int BUFFERSIZE_64 = 64;
-constexpr unsigned int BUFFERSIZE_128 = 128;
-constexpr unsigned int BUFFERSIZE_256 = 256;
+constexpr size_t BUFFERSIZE_64 = 64;
+constexpr size_t BUFFERSIZE_128 = 128;
+constexpr size_t BUFFERSIZE_256 = 256;
 
 enum class BufferFormat {
     Mono,
